+++ logout.php
@@ -12,7 +12,7 @@
 	</title>
 
 	<head>
-		<?php require 'php/css.php'; ?>
+		<link rel="stylesheet" type="text/css" href="login.css">
 		<script type="text/javascript" language="JavaScript">
 				setTimeout(function() {window.location = 'login.php'}, 1000);
 		</script>