+++ setsettings.php
@@ -1,7 +1,9 @@
 <?php
 	require_once('auth.php');
 	require "php/mysqlconnect.php";
+	require 'php/rights.php';
 	
+	if ("${UserType[UserType]}" == "admin") {
 	//Global Theme
 		$NewTheme = $_GET['theme'];
 	
@@ -18,4 +20,22 @@
 		echo "	<script type=\"text/javascript\" language=\"JavaScript\">
 			setTimeout(function() {window.location = 'settings.php?page=general'}, 0);
              	</script>";
+	} else {
+	//User Theme
+		$NewTheme = $_GET['theme'];
+	
+		$sql = $DBH->prepare("SELECT parameter FROM settings WHERE setting='theme' AND user='$LoggedInUserName'");
+		$sql->execute();
+		$themeResult = $sql->fetch();
+		$CurrentTheme = "${themeResult[parameter]}";
+	
+		if ($CurrentTheme !== $NewTheme) {
+			$sql = $DBH->prepare("UPDATE settings SET parameter='$NewTheme' WHERE setting='theme' AND user='$LoggedInUserName'");
+			$sql->execute();
+		}
+		
+		echo "	<script type=\"text/javascript\" language=\"JavaScript\">
+			setTimeout(function() {window.location = 'settings.php?page=personal'}, 0);
+             	</script>";
+	}
 ?>